requirements.txt
aiohttp==3.7.3
aioredis==1.3.1
aniso8601==8.1.0
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
coinmarketcap
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.5
django-allauth==0.44.0
django-appconf==1.0.4
django-colorfield
django-countries==7.0
django-defender==0.8.0
django-ipware==3.0.2
django-otp==1.1.3
django-prometheus==2.1.0
django-redis==4.12.1
django-rest-auth==0.9.5
django-tracking-analyzer==1.1.1
django-user-agents==0.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-spectacular==0.14.0
drf-writable-nested==0.6.2
drf-yasg==1.20.0
dry-rest-permissions==0.1.10
geoip2==4.1.0
gunicorn==20.0.4
hiredis==1.1.0
html5lib==1.0.1
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
inflection==0.5.1
ipaddr==2.2.0
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
JSON-log-formatter==0.3.1
jsonschema==3.2.0
lockfile==0.12.2
Mako==1.1.4
MarkupSafe==1.1.1
maxminddb==2.0.3
msgpack==1.0.2
multidict==5.1.0
mypy-extensions==0.4.3
oauthlib==3.1.0
opentracing==2.4.0
packaging==20.8
pathspec==0.8.1
pep517==0.8.2
Pillow==8.1.0
progress==1.5
prometheus-client==0.11.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.5
PyYAML==5.4.1
redis==3.5.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rest-framework-cache==0.1
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
threadloop==1.0.2
thrift==0.13.0
toml==0.10.2
tornado==6.1
Twisted==20.3.0
txaio==21.2.1
typing-extensions==3.7.4.3
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.2
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
yarl==1.6.3
zipp==3.4.0
zope.interface==5.2.0
